+++ packages/postcss-plugin-transform-web-tag/script/build.js
@@ -2,7 +2,7 @@ const { execCommand } = require('t-comm');
 const path = require('path');
 
 function main() {
-  const source = path.resolve(__dirname, '../src/index.js');
+  const source = path.resolve(__dirname, '../src/*');
   const target = path.resolve(__dirname, '../lib/');
   console.log('[source]', source);
   console.log('[target]', target);