+++ packages/postcss-plugin-transform-web-tag/src/index.d.ts
@@ -1,0 +1,4 @@
+export interface Options {
+  tagMap?: Record<string, string>;
+}
+export default function transformWebTag(options?: undefined | Options): void;