+++ packages/project-config-uni-vite/src/types.ts
@@ -1,7 +1,12 @@
 import type { Plugin, ServerOptions, CommonServerOptions } from 'vite';
-import type { IRemoveVueDirectionOptions } from '@plugin-light/vite-plugin-remove-vue-directive';
 import type { Server } from 'node:https';
+import tailwindcss from 'tailwindcss';
+import type { UniTailwindPluginUserOptions } from '@uni-helper/vite-plugin-uni-tailwind';
+
+import type { IRemoveVueDirectionOptions } from '@plugin-light/vite-plugin-remove-vue-directive';
 import type { ICrossGameStyleOptions } from '@plugin-light/vite-plugin-cross-game-style';
+import type { Options as TransformWebTagOptions } from '@plugin-light/postcss-plugin-transform-web-tag';
+
 
 export interface IUniViteConfigOptions {
   mode: string;
@@ -28,4 +33,13 @@ export interface IUniViteConfigOptions {
 
   // 语法警报列表，比如 v-model，destroyed
   warnList?: ICrossGameStyleOptions['warnList'];
+
+  // transform-web-tag postcss插件
+  transformWebTagOptions?: boolean | TransformWebTagOptions;
+
+  // uniTailwind 插件
+  uniTailwindOptions?: boolean | UniTailwindPluginUserOptions;
+
+  // tailwindcss postcss插件
+  tailwindcssOptions?: boolean | Parameters<typeof tailwindcss>[0];
 }