+++ packages/plugin-light-shared/CHANGELOG.md
@@ -1,3 +1,14 @@
+## <small>0.0.12 (2024-09-29)</small>
+
+* feat(tailwind): add webpack-plugin-tailwind ([444a3ad](https://github.com/novlan1/plugin-light/commits/444a3ad))
+* chore: update name ([02b2100](https://github.com/novlan1/plugin-light/commits/02b2100))
+* chore: update tsconfig.json ([7d20e75](https://github.com/novlan1/plugin-light/commits/7d20e75))
+* test: update packages tests ([6276643](https://github.com/novlan1/plugin-light/commits/6276643))
+* test: update packages tests ([4897b59](https://github.com/novlan1/plugin-light/commits/4897b59))
+* docs: update docs ([883589c](https://github.com/novlan1/plugin-light/commits/883589c))
+
+
+
 ## <small>0.0.11 (2024-09-19)</small>
 
 * chore: add meta in pkg.json ([3f6088c](https://github.com/novlan1/plugin-light/commits/3f6088c))