+++ packages/postcss-plugin-transform-web-tag/CHANGELOG.md
@@ -1,3 +1,10 @@
+## <small>0.0.3 (2024-09-30)</small>
+
+* feat(uni-vite): support tailwindcss ([e915dcc](https://github.com/novlan1/plugin-light/commits/e915dcc))
+* chore: update plugin-light-shared to 0.0.12 ([4d33550](https://github.com/novlan1/plugin-light/commits/4d33550))
+
+
+
 ## <small>0.0.2 (2024-09-25)</small>
 
 * feat(postcss-plugin): add postcss plugin transform-web-tag ([89e83fc](https://github.com/novlan1/plugin-light/commits/89e83fc))