+++ packages/project-config-uni-vite/CHANGELOG.md
@@ -1,3 +1,20 @@
+## <small>0.0.11 (2024-09-30)</small>
+
+* chore: add meta in pkg.json ([3f6088c](https://github.com/novlan1/plugin-light/commits/3f6088c))
+* chore: add release script ([defe6af](https://github.com/novlan1/plugin-light/commits/defe6af))
+* chore: update deps to latest ([455918d](https://github.com/novlan1/plugin-light/commits/455918d))
+* chore: update name ([02b2100](https://github.com/novlan1/plugin-light/commits/02b2100))
+* chore: update plugin-light-shared ([08a9790](https://github.com/novlan1/plugin-light/commits/08a9790))
+* chore: update plugin-light-shared to 0.0.12 ([4d33550](https://github.com/novlan1/plugin-light/commits/4d33550))
+* chore: update tsconfig.json ([7d20e75](https://github.com/novlan1/plugin-light/commits/7d20e75))
+* chore(uni-vite): update postcss-plugin-transfrom-web-tag ([dd26a4d](https://github.com/novlan1/plugin-light/commits/dd26a4d))
+* feat(uni-vite): support tailwindcss ([e915dcc](https://github.com/novlan1/plugin-light/commits/e915dcc))
+* test: update packages tests ([4897b59](https://github.com/novlan1/plugin-light/commits/4897b59))
+* docs: move docs position ([1937e17](https://github.com/novlan1/plugin-light/commits/1937e17))
+* docs: update docs ([883589c](https://github.com/novlan1/plugin-light/commits/883589c))
+
+
+
 ## <small>0.0.10 (2024-09-18)</small>
 
 * chore: update project-config ([af4ff17](https://github.com/novlan1/plugin-light/commits/af4ff17))